refactor: extract map runner episode execution
